fix(headless): extract write-path scanning to _headless_scan.py to meet line budget
headless.py grew to 1120 lines (exemption: 1100) after the Channel B
DIR_MISSING detection code was added.  Extract the write-path JSONL scanning
logic (_scan_jsonl_write_paths + 4 private constants) to a new stdlib-only
module src/autoskillit/execution/_headless_scan.py.  headless.py re-imports
the function so all existing import paths remain valid.  Net reduction:
1120 → 1045 lines.  Update REQ-CNST-003-E3 execution/ file-count exemption
(25 → 26) with rationale, and register _headless_scan.py in CLAUDE.md.
